#  Reduced memory allocation by skipping string-allocation