--- a/src/NLog.StructuredLogging.Json/FlattenedJsonLayout.cs
+++ b/src/NLog.StructuredLogging.Json/FlattenedJsonLayout.cs
@@ -6,13 +6,17 @@ using NLog.Config;
 using NLog.StructuredLogging.Json.Helpers;
 using NLog.LayoutRenderers.Wrappers;
 using NLog.Layouts;
+using Newtonsoft.Json;
 
 namespace NLog.StructuredLogging.Json
 {
     [Layout("FlattenedJsonLayout")]
-    [AppDomainFixedOutput]
+    [ThreadSafe]
     public class FlattenedJsonLayout : JsonLayout
     {
+        private JsonSerializer JsonSerializer => _jsonSerializer ?? (_jsonSerializer = ConvertJson.CreateJsonSerializer());
+        private JsonSerializer _jsonSerializer;
+
         public FlattenedJsonLayout()
         {
             SuppressSpaces = true;
@@ -26,26 +30,53 @@ namespace NLog.StructuredLogging.Json
 
         protected override string GetFormattedMessage(LogEventInfo logEvent)
         {
-            var result = new Dictionary<string, object>();
-
-            AppendDataFromAttributes(logEvent, result);
-            AppendLogProperties(logEvent, result);
-            AppendLogParameters(logEvent, result);
-            AppendExceptionData(logEvent, result);
-
+            var result = BuildPropertiesDictionaryFlattened(logEvent);
             return ConvertJson.Serialize(result);
         }
 
         protected override void RenderFormattedMessage(LogEventInfo logEvent, StringBuilder target)
         {
-            target.Append(GetFormattedMessage(logEvent) ?? string.Empty);
+            var result = BuildPropertiesDictionaryFlattened(logEvent);
+            var orgLength = target.Length;
+
+            try
+            {
+                // Ensure we are threadsafe
+                var jsonSerializer = JsonSerializer;
+                lock (jsonSerializer)
+                {
+                    // Serialize directly into StringBuilder
+                    ConvertJson.Serialize(result, target, jsonSerializer);
+                }
+            }
+            catch
+            {
+                _jsonSerializer = null; // Do not reuse JsonSerializer, as it might have become broken
+                target.Length = orgLength;  // Skip invalid JSON
+                throw;
+            }
+        }
+
+        private Dictionary<string, object> BuildPropertiesDictionaryFlattened(LogEventInfo logEvent)
+        {
+            var result = new Dictionary<string, object>();
+            AppendDataFromAttributes(logEvent, result);
+            AppendLogProperties(logEvent, result);
+            AppendLogParameters(logEvent, result);
+            AppendExceptionData(logEvent, result);
+            return result;
         }
 
         private void AppendDataFromAttributes(LogEventInfo logEvent, IDictionary<string, object> result)
         {
+            if (Attributes.Count == 0)
+                return;
+
             var layoutRendererWrapper = new JsonEncodeLayoutRendererWrapper();
-            foreach (var jsonAttribute in Attributes)
+            // Enumerate without allocation of GetEnumerator()
+            for (int i = 0; i < Attributes.Count; ++i)
             {
+                var jsonAttribute = Attributes[i];
                 AddRenderedValue(logEvent, result, layoutRendererWrapper, jsonAttribute);
             }
         }
@@ -96,29 +127,27 @@ namespace NLog.StructuredLogging.Json
 
         private static void AppendExceptionData(LogEventInfo logEvent, IDictionary<string, object> result)
         {
-            if (logEvent.Exception == null)
+            if (logEvent.Exception?.Data?.Count > 0)
             {
-                return;
+                Mapper.HarvestToDictionary(logEvent.Exception.Data, result, "ex_");
             }
-
-            Mapper.HarvestToDictionary(logEvent.Exception.Data, result, "ex_");
         }
 
         private static void AppendLogProperties(LogEventInfo logEvent, IDictionary<string, object> result)
         {
-            Mapper.HarvestToDictionary(logEvent.Properties, result, "data_");
+            if (logEvent.HasProperties)
+            {
+                Mapper.HarvestToDictionary(logEvent.Properties, result, "data_");
+            }
         }
 
         private static void AppendLogParameters(LogEventInfo logEvent, IDictionary<string, object> result)
         {
-            if (logEvent.Parameters == null)
+            if (logEvent.Parameters != null)
             {
-                return;
+                var value = string.Join(",", logEvent.Parameters.Select(x => x ?? "null").Select(x => x.ToString()));
+                result.Add("Parameters", value);
             }
-
-            var value = string.Join(",", logEvent.Parameters.Select(x => x ?? "null").Select(x => x.ToString()));
-
-            result.Add("Parameters", value);
         }
 
         private void Add(string name, Layout layout, bool encode = false)