--- a/src/NLog.StructuredLogging.Json/Helpers/ConvertJson.cs
+++ b/src/NLog.StructuredLogging.Json/Helpers/ConvertJson.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+using System.Collections.Generic;
+using System.IO;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 
@@ -6,15 +7,28 @@ namespace NLog.StructuredLogging.Json.Helpers
 {
     public static class ConvertJson
     {
-        private static readonly JsonSerializerSettings LogSettings = new JsonSerializerSettings
+        internal static readonly JsonSerializerSettings LogSettings = new JsonSerializerSettings
         {
             ContractResolver = new DefaultContractResolver(),
             Formatting = Formatting.None
         };
 
+        internal static JsonSerializer CreateJsonSerializer() => JsonSerializer.CreateDefault(LogSettings);
+
         public static string Serialize(Dictionary<string, object> data)
         {
             return JsonConvert.SerializeObject(data, LogSettings);
+        }
+
+        public static void Serialize(Dictionary<string, object> data, System.Text.StringBuilder sb, JsonSerializer jsonSerializer)
+        {
+            using (var sw = new StringWriter(sb))
+            {
+                using (JsonWriter writer = new JsonTextWriter(sw))
+                {
+                    jsonSerializer.Serialize(writer, data);
+                }
+            }
         }
     }
 }