--- a/src/NLog.StructuredLogging.Json/JsonWithPropertiesLayout.cs
+++ b/src/NLog.StructuredLogging.Json/JsonWithPropertiesLayout.cs
@@ -1,5 +1,7 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
+using System.Text;
+using Newtonsoft.Json;
 using NLog.Config;
 using NLog.Layouts;
 using NLog.StructuredLogging.Json.Helpers;
@@ -26,8 +28,12 @@ namespace NLog.StructuredLogging.Json
     ///   </target>
     /// </example>
     [Layout("jsonwithproperties")]
+    [ThreadSafe]
     public class JsonWithPropertiesLayout : Layout
     {
+        private JsonSerializer JsonSerializer => _jsonSerializer ?? (_jsonSerializer = ConvertJson.CreateJsonSerializer());
+        private JsonSerializer _jsonSerializer;
+
         [ArrayParameter(typeof(StructuredLoggingProperty), "property")]
         public IList<StructuredLoggingProperty> Properties { get; private set; }
 
@@ -40,14 +46,45 @@ namespace NLog.StructuredLogging.Json
 
         protected override string GetFormattedMessage(LogEventInfo logEvent)
         {
+            Dictionary<string, object> dictionary = BuildPropertiesDictionary(logEvent);
+            return ConvertJson.Serialize(dictionary);
+        }
+
+        protected override void RenderFormattedMessage(LogEventInfo logEvent, StringBuilder target)
+        {
+            var result = BuildPropertiesDictionary(logEvent);
+            var orgLength = target.Length;
+
+            try
+            {
+                // Ensure we are threadsafe
+                var jsonSerializer = JsonSerializer;
+                lock (jsonSerializer)
+                {
+                    // Serialize directly into StringBuilder
+                    ConvertJson.Serialize(result, target, jsonSerializer);
+                }
+            }
+            catch
+            {
+                _jsonSerializer = null; // Do not reuse JsonSerializer, as it might have become broken
+                target.Length = orgLength;  // Skip invalid JSON
+                throw;
+            }
+        }
+
+        private Dictionary<string, object> BuildPropertiesDictionary(LogEventInfo logEvent)
+        {
             var dictionary = Mapper.ToDictionary(logEvent);
 
-            foreach (var property in Properties)
+            // Enumerate without allocation of GetEnumerator()
+            for (int i = 0; i < Properties.Count; ++i)
             {
+                var property = Properties[i];
                 AddRenderedValue(logEvent, dictionary, property);
             }
 
-            return ConvertJson.Serialize(dictionary);
+            return dictionary;
         }
 
         private static void AddRenderedValue(