Reduced memory allocation by skipping string-allocation by writing directly to StringBuilder and reusing JsonSerializer. Improved concurrency by marking Layout as ThreadSafe.

Performance improved by 10 pct. Memory Allocation reduced by 30 pct.