@snakefoot 2019-08-20
**Benchmark after:**

|                        Method |       Mean |     Error |     StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------ |-----------:|----------:|-----------:|-------:|------:|------:|----------:|
|                       LogInfo |   3.517 us | 0.0346 us |  0.0323 us | 0.5646 |     - |     - |   1.74 KB |
|               LogExtendedInfo |  22.217 us | 0.3460 us |  0.3236 us | 2.9602 |     - |     - |   9.17 KB |
| LogExtendedInfoWithProperties |  27.461 us | 0.3291 us |  0.3078 us | 3.4790 |     - |     - |  10.76 KB |
|                  LogException |   8.595 us | 0.0599 us |  0.0560 us | 1.2817 |     - |     - |   3.94 KB |
|          LogExtendedException |  32.568 us | 0.0978 us |  0.0816 us | 3.8452 |     - |     - |  11.88 KB |
|    LogExceptionWithProperties | 111.220 us | 6.8614 us | 20.1234 us |      - |     - |     - |  15.32 KB |

**Benchmark before:**

|                        Method |       Mean |     Error |     StdDev |     Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------ |-----------:|----------:|-----------:|-----------:|-------:|------:|------:|----------:|
|                       LogInfo |   3.792 us | 0.0344 us |  0.0305 us |   3.790 us | 0.9308 |     - |     - |   2.86 KB |
|               LogExtendedInfo |  24.512 us | 0.1101 us |  0.0976 us |  24.544 us | 3.3569 |     - |     - |   10.4 KB |
| LogExtendedInfoWithProperties |  29.095 us | 0.3760 us |  0.3517 us |  29.169 us | 4.1199 |     - |     - |   12.7 KB |
|                  LogException |   9.296 us | 0.0448 us |  0.0419 us |   9.309 us | 1.9836 |     - |     - |    6.1 KB |
|          LogExtendedException |  34.988 us | 0.4165 us |  0.3896 us |  35.099 us | 4.5776 |     - |     - |  14.19 KB |
|    LogExceptionWithProperties | 108.411 us | 3.9489 us | 11.5191 us | 105.279 us |      - |     - |     - |  18.88 KB 

BenchmarkDotNet=v0.11.5, OS=Windows 10.0.17134.950 (1803/April2018Update/Redstone4)
Intel Core i5-4460 CPU 3.20GHz (Haswell), 1 CPU, 4 logical and 4 physical cores
Frequency=3125015 Hz, Resolution=319.9985 ns, Timer=TSC
.NET Core SDK=2.1.508
  [Host]     : .NET Core 2.1.12 (CoreCLR 4.6.27817.01, CoreFX 4.6.27818.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.1.12 (CoreCLR 4.6.27817.01, CoreFX 4.6.27818.01), 64bit RyuJIT

** Legends **
  Mean      : Arithmetic mean of all measurements
  Error     : Half of 99.9% confidence interval
  StdDev    : Standard deviation of all measurements
  Gen 0     : GC Generation 0 collects per 1000 operations
  Gen 1     : GC Generation 1 collects per 1000 operations
  Gen 2     : GC Generation 2 collects per 1000 operations
  Allocated : Allocated memory per single operation (managed only, inclusive, 1KB = 1024B)
  1 us      : 1 Microsecond (0.000001 sec)